Implement the add opcode.
The add opcode

The opcode add adds the top two elements of the stack.

Usage: add
If the stack contains less than two elements, print the error message L<line_number>: can't add, stack too short, followed by a new line, and exit with the status EXIT_FAILURE
The result is stored in the second top element of the stack, and the top element is removed, so that at the end:
The top element of the stack contains the result
The stack is one element shorter

+++ monty_file_reader_algo.c
@@ -52,7 +52,7 @@ void push_to_stack_algo(char *str, int i, int x,
 		fprintf(stderr, "L%d: usage: push integer\n", line_number);
 		exit(EXIT_FAILURE);
 	}
-	opcode_function_caller("push", &(*stack), (unsigned int) atoi(code));
+	opcode_function_caller("push", &(*stack), atoi(code));
 }
 
 /**
